pyproject.toml
[tool.poetry]
name = "uchprak"
version = "0.1.0"
description = ""
authors = ["LetsAllLoveLain <qwspam3@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
psycopg2 = "^2.9.10"
sqlalchemy = "^2.0.36"
pydantic = "^2.10.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
